requirements.txt
flask
flask-cors
flask-sqlalchemy
mysqlclient
pyjwt
python-dotenv
